% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Question dictionary}
\usage{
dictionary(code)
}
\arguments{
\item{code}{character, the code corresponding to the question.}
}
\description{
Displays information from the questions in the "Your DNA, Your Say" survey.
}
\examples{
dictionary("q_1.1.1a")

}
